Move physical device info to rhi
- Change casing for values in physical device type enum
- Change order of dependencies in editor and demo
