+++ demos/pong-3d/premake5.lua
@@ -3,7 +3,7 @@ project "DemoPong"
     kind "ConsoleApp"
 
     loadSourceFiles{}
-    linkDependenciesWith{"LiquidEngineRHIVulkan", "LiquidEngine", "LiquidEngineRHICore"}
+    linkDependenciesWith{"LiquidEngine", "LiquidEngineRHIVulkan", "LiquidEngineRHICore"}
 
     postbuildcommands {
         "glslc ../../../demos/pong-3d/assets/basic-shader.vert -o %{cfg.buildtarget.directory}/basic-shader.vert.spv",
