+++ engine/tests/liquid-tests/profiler/PhysicalDeviceInformation.test.cpp
@@ -1,17 +1,17 @@
 #include "liquid/core/Base.h"
-#include "liquid/profiler/PhysicalDeviceInformation.h"
+#include "liquid/rhi/PhysicalDeviceInformation.h"
 
 #include "liquid-tests/Testing.h"
 
 TEST(PhysicalDeviceInformationTest, SetsNameTypePropertiesAndLimits) {
-  liquid::PhysicalDeviceInformation info(
-      "Test Device", liquid::PhysicalDeviceType::DISCRETE_GPU,
+  liquid::rhi::PhysicalDeviceInformation info(
+      "Test Device", liquid::rhi::PhysicalDeviceType::DiscreteGPU,
       {{"driverVersion", liquid::Property(120u)},
        {"apiVersion", liquid::Property(1u)}},
       {{"maxViewports", liquid::Property(1u)}});
 
   EXPECT_EQ(info.getName(), "Test Device");
-  EXPECT_EQ(info.getType(), liquid::PhysicalDeviceType::DISCRETE_GPU);
+  EXPECT_EQ(info.getType(), liquid::rhi::PhysicalDeviceType::DiscreteGPU);
 
   EXPECT_EQ(info.getProperties().size(), 2);
   EXPECT_EQ(info.getProperties()[0].first, "driverVersion");
